+++ src/index.ts
@@ -1,3 +1,4 @@
+export { batch } from './batch';
 export { computed, type Computed } from './computed';
 export {
   effect,